% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_ls}
\alias{aws_ls}
\title{List all objects in an AWS bucket}
\usage{
aws_ls(bucket = "publicaccountability", prefix = "csv", ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket, or an object of
class “s3_bucket”.}

\item{prefix}{Character string that limits the response to keys that begin
with the specified prefix.}

\item{...}{Additional arguments passed to \code{\link[aws.s3:s3HTTP]{aws.s3::s3HTTP()}}.}
}
\description{
Uses \code{\link[aws.s3:get_bucket]{aws.s3::get_bucket_df()}} styled as an 'fs' data frame.
}
