% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check2.R
\name{check_for_cran2}
\alias{check_for_cran2}
\title{Check an R-package on R-hub, for a CRAN submission}
\usage{
check_for_cran2(path = ".", email = NULL)
}
\arguments{
\item{path}{Path to a directory containing an R package, or path to source R
package tarball built with R CMD build or \code{devtools::build()}.}

\item{email}{Email address to send notification to about the check. It must
be a validated email address, see validate_email(). If NULL, then the email
address of the maintainer is used, as defined in the DESCRIPTION file of
the package.}
}
\description{
Per issue \href{https://github.com/r-hub/rhub/issues/367}{r-hub/rhub#367}, this
calls \code{\link[rhub:check_for_cran]{rhub::check_for_cran()}} with the \code{R_COMPILE_AND_INSTALL_PACKAGES}
flag set to "always" to avoid package installation issues.
}
