% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\name{copy_last}
\alias{copy_last}
\title{Copy the last object to the clipboard}
\usage{
copy_last(x = .Last.value)
}
\arguments{
\item{x}{The object to view, usually left as \link[base:Last.value]{base::.Last.value}.}
}
\value{
The same \code{.Last.value} as \emph{before} copied, invisibly.
}
\description{
Use \code{\link[clipr:write_clip]{clipr::write_clip()}} to write the last value as a character vector to
the system clipboard.
}
\details{
The value of the internal evaluation of a top-level R expression is always
assigned to \code{.Last.value} before further processing (e.g., printing).
}
