% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count.character}
\alias{count.character}
\title{Count values in a character vector}
\usage{
count.character(x, sort = TRUE, prop = TRUE)
}
\arguments{
\item{x}{A character vector.}

\item{sort}{If \code{TRUE}, will show the largest groups at the top.}

\item{prop}{If \code{TRUE}, compute the fraction of marginal table.}
}
\value{
A tibble of element counts
}
\description{
Method for \code{\link[dplyr:count]{dplyr::count()}} (or \code{\link[=count]{count()}} wrapper).
}
\examples{
x <- sample(LETTERS)[rpois(1000, 10)]
count(x)
}
