% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count}
\alias{count}
\title{Count the way you want}
\usage{
count(x, ..., wt = NULL, sort = TRUE, prop = TRUE)
}
\arguments{
\item{x}{A data frame.}

\item{...}{Variables to group by.}

\item{wt}{Frequency weights.}

\item{sort}{If \code{TRUE}, will show the largest groups at the top.}

\item{prop}{If \code{TRUE}, compute the fraction of marginal table.}
}
\value{
A tibble of element counts
}
\description{
A wrapper around \code{\link[dplyr:count]{dplyr::count()}} with \code{sort} set to \code{TRUE} by default and
the an additional column created by default containing the proportional
fraction each observation makes of the whole.
}
\examples{
count(iris, Species)
}
