% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnd-dpr.R
\name{dnd_dpr}
\alias{dnd_dpr}
\title{Damage Per Round}
\usage{
dnd_dpr(M, A, D, B, C = 20, adv = NA)
}
\arguments{
\item{M}{Monster}

\item{A}{Attack: Attack bonus (Proficiency + STR/DEX).}

\item{D}{Damage: Average dice roll (e.g., d8 = 4.5).}

\item{B}{Bonus: STR/DEX.}

\item{C}{Critical: Critical hit rolls (Usually 20, sometimes lower).}

\item{adv}{Advantage: \code{NA}, \code{TRUE}, or \code{FALSE.}}
}
\description{
Damage Per Round
}
\examples{
dnd_dpr(15, 8, 4.5, 5)
}
\seealso{
\url{https://rpgbot.net/dnd5/characters/damage-per-round.html}
}
