% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_rx}
\alias{filter_rx}
\title{Filter a data frame by a regular expression}
\usage{
filter_rx(dat, col, pattern, ...)
}
\arguments{
\item{dat}{A data frame with a character column to filter.}

\item{col}{The column containing a character vector to input.}

\item{pattern}{Pattern to look for..}

\item{...}{Additional arguments passed to \code{\link[stringr:str_detect]{stringr::str_detect()}}.}
}
\value{
A subset of rows from \code{dat}.
}
\description{
A shortcut for \code{dat \%>\% filter(str_detect(column, "\\\\d"))}.
}
