% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaa}
\alias{gaa}
\title{GAA Team Abbreviations by Season and Team ID}
\format{
A data frame with 74 rows and 3 variables:
\describe{
\item{seasonId}{The fantasy season integer}
\item{teamId}{The team ID integer}
\item{abbrev}{The normalized \emph{owner} abbreviation for that year}
...
}
}
\usage{
gaa
}
\description{
GAA Team Abbreviations by Season and Team ID
}
\keyword{datasets}
