% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\name{last_value}
\alias{last_value}
\title{Return the last value}
\usage{
last_value(x = .Last.value)
}
\description{
A function shortcut for accessing \link{.Last.value}.
}
\details{
The value of the internal evaluation of a top-level R expression is always
assigned to .Last.value (in package:base) before further processing (e.g.,
printing).
}
