% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{load.packages}
\alias{load.packages}
\alias{save.packages}
\title{Save and load packages from file}
\usage{
load.packages(path = NULL, install = FALSE)

save.packages(x = NULL, path = tempfile())
}
\arguments{
\item{path}{The path to a text file containing package names. If \code{NULL}
(default), then the default list is read from \code{k5/inst/PACKAGES}.}

\item{install}{If \code{TRUE}, install missing packages.}

\item{x}{A character vector of package names to save. If \code{NULL} (default),
use all currently attached packages.}
}
\value{
The list of packages, invisibly.
}
\description{
Save and load packages from file
}
