% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{load_my_packages}
\alias{load_my_packages}
\alias{save_my_packages}
\title{Save and load packages from file}
\usage{
load_my_packages(path = NULL)

save_my_packages(x = NULL, path = tempfile())
}
\arguments{
\item{path}{The path to a text file containing one package per line. If
\code{NULL} (default), then the default list is read from \code{k5/inst/PACKAGES}.}

\item{x}{A character vector of package names to save. If \code{NULL} (default),
use all currently attached packages.}
}
\value{
The list of packages, invisibly.
}
\description{
Save and load packages from file
}
