% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_csv_dumb}
\alias{read_csv_dumb}
\title{Read a CSV file without column guessing}
\usage{
read_csv_dumb(file, ...)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data.}

\item{...}{Additional arguments passed to \code{\link[readr:read_delim]{readr::read_csv()}}.}
}
\description{
Use \code{\link[readr:read_delim]{readr::read_csv()}} without specifying \emph{any} column types. All columns
are treated as character strings.
}
