% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_delim_aws}
\alias{read_delim_aws}
\title{Read a CSV file from AWS object}
\usage{
read_delim_aws(object, bucket = aws_bucket(), delim = ",", ...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[readr:read_delim]{readr::read_delim()}}.}
}
\value{
A data frame read from a remote file.
}
\description{
Use \code{\link[readr:read_delim]{readr::read_delim()}} on a comma-delimited text file stored as an object
on an AWS S3 bucket.
}
