% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_delim_clip}
\alias{read_delim_clip}
\title{Read a table from the clipboard}
\usage{
read_delim_clip(delim = "\\t", ...)
}
\arguments{
\item{delim}{Single character used to separate fields within a record.}

\item{...}{Additional arguments passed to \code{\link[readr:read_delim]{readr::read_delim()}}.}
}
\value{
A data frame read from the clipboard.
}
\description{
Use \code{\link[readr:read_delim]{readr::read_delim()}} on a string copied to the clipboard. Defaults to
tab separator like given when copying cells from spreadsheets.
}
