% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_delim_dumb}
\alias{read_delim_dumb}
\alias{read_csv_dumb}
\alias{read_tsv_dumb}
\title{Read a text file without column guessing}
\usage{
read_delim_dumb(file, delim = c(",", "\\t", "|"), ...)

read_csv_dumb(file, ...)

read_tsv_dumb(file, ...)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data.}

\item{delim}{Single character used to separate fields within a record.}

\item{...}{Additional arguments passed to \code{\link[readr:read_delim]{readr::read_delim()}}.}
}
\value{
A tibble data frame read from the file.
}
\description{
Use \code{\link[readr:read_delim]{readr::read_delim()}} without specifying \emph{any} column types. All columns
are treated as character strings.
}
