% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar.R
\name{tar_create}
\alias{tar_create}
\title{Create tar archive}
\usage{
tar_create(
  path,
  tar = NULL,
  compress = c("xz", "none", "gzip", "bzip2"),
  level = 6,
  extra = "",
  method = "tar"
)
}
\arguments{
\item{path}{A character vector of files to archive.}

\item{tar}{The path of the tarball file to create. If none is supplied, and
only one path is, an archive with that \code{\link[=basename]{basename()}} will be created in the
same directory.}

\item{compress}{Character string giving the type of compression to be used.}

\item{level}{Integer compression level used with internal method, 1-9.}

\item{extra}{An optional character vector passed to \code{\link[=system]{system()}} as arguments
of the external \code{tar} command.}

\item{method}{A character string specifying the external command to be used.}
}
\value{
The path of the created archive (invisibly).
}
\description{
A wrapper for an external \code{tar} command to create tar archives. Unlike
\code{\link[utils:tar]{utils::tar()}}, filepaths can be passed into the first argument with \verb{\%>\%}.
}
\examples{
\dontshow{.old_wd <- setwd(tempdir())}
# Tar a file to same basename
tmp <- tempfile(fileext = ".csv")
write.csv(iris, tmp)
(txz <- tar_create(tmp))
file.size(c(tmp, txz))
\dontshow{setwd(.old_wd)}
}
