% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-stats.R
\name{var_missing}
\alias{var_missing}
\alias{var_distinct}
\title{Apply a statistic function to all variables}
\usage{
var_missing(df)

var_distinct(df)
}
\arguments{
\item{df}{A data frame to glimpse.}
}
\value{
Invisibly, a table of statistics by column of a data frame.
}
\description{
Apply either \code{\link[=count_na]{count_na()}} or \code{\link[dplyr:n_distinct]{dplyr::n_distinct()}} to every column of a data
frame and return the count and share of total values (either proportion
missing or proportion distinct).
}
\examples{
var_missing(dplyr::storms)
var_distinct(dplyr::storms)
}
