% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-firefox.R
\name{view_firefox}
\alias{view_firefox}
\title{View an HTML document in Firefox}
\usage{
view_firefox(html)
}
\arguments{
\item{html}{An object which has the class \code{xml_document}, often from rvest.}
}
\value{
The html object, invisibly.
}
\description{
Take an XML document object, write to an HTML file, and open in Firefox.
}
