% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\name{view_last}
\alias{view_last}
\title{View the last object}
\usage{
view_last(x = .Last.value)
}
\arguments{
\item{x}{The object to view, usually left as \link[base:Last.value]{base::.Last.value}.}
}
\value{
The same \code{.Last.value} as \emph{before} viewing, invisibly.
}
\description{
Invoke a spreadsheet-style data viewer on a matrix-like R object. In a
non-interactive session, the object is returned invisibly and nothing is
"viewed".
}
\details{
The value of the internal evaluation of a top-level R expression is always
assigned to \code{.Last.value} before further processing (e.g., printing).
}
