% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-count.R
\name{word_count}
\alias{word_count}
\title{Count file words, lines, and bytes}
\usage{
word_count(path, count = "")
}
\arguments{
\item{path}{Character vector of file paths.}

\item{count}{The type of element to count, see details.}
}
\value{
A data frame of counts by file.
}
\description{
Invoke system tool to print newline, word, and byte counts for each file.
}
\details{
One of five options or an empty string (default):
\enumerate{
\item "lines" for newline characters (separating lines).
\item "words" for words separated by white space.
\item "chars" for individual characters.
\item "bytes" for total bytes, differs with multibyte characters.
\item "max" for the maximum display width of longest line.
}
}
