% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{write_delim_clip}
\alias{write_delim_clip}
\title{Write a table from the clipboard}
\usage{
write_delim_clip(x, delim = "\\t", ...)
}
\arguments{
\item{x}{A data frame to write to clipboard.}

\item{delim}{Single character used to separate fields within a record.}

\item{...}{Additional arguments passed to \code{\link[readr:format_delim]{readr::format_delim()}}.}
}
\value{
Invisibly, the input data frame.
}
\description{
Use \code{\link[readr:format_delim]{readr::format_delim()}} on a data frame to copy a string to the
clipboard. Defaults to tab separator like given when copying cells from
spreadsheets.
}
