% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{zip_example}
\alias{zip_example}
\title{Get path to k5 example}
\usage{
zip_example(zip = TRUE)
}
\arguments{
\item{zip}{If \code{TRUE} (default), the path to the compressed version is
returned. If \code{FALSE}, the uncompressed version. If \code{NULL}, the example
file names (not paths) will be listed.}
}
\value{
If \code{zip} is not \code{NULL}, the \code{fs_path} to the system file.
}
\description{
k5 comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access. The two files are regular
and compressed versions of a delimited file containing the mtcars data.
}
\examples{
zip_example()
zip_example(NULL)
}
