% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_all}
\alias{print_all}
\title{Print All Rows}
\usage{
print_all(x)
}
\arguments{
\item{x}{Object to print, typically a data frame or vector.}
}
\description{
If the object is a tibble, all rows are printed.
}
