% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{zip_info}
\alias{zip_info}
\alias{zip_ls}
\title{List zip contents}
\usage{
zip_info(path, method = "internal")

zip_ls(path, method = "internal")
}
\arguments{
\item{path}{A character vector of zip archives.}

\item{method}{The method to be used. Passed to \code{\link[utils:unizp]{utils::unizp()}}. An
alternative is to use \code{getOption("unzip")}.}
}
\description{
\code{\link[=zip_info]{zip_info()}} is the equivalent to the \code{unzip -l} command. It returns a data
frame (like \code{\link[fs:dir_ls]{fs::dir_info()}}).

\code{\link[=zip_ls]{zip_ls()}} also uses \code{unzip -l} but only returns the filenames within the
archive as a named \code{fs_path} character vector (like \code{\link[=fs:dir_ls]{fs:dir_ls()}}).
}
\examples{
# List two files from zip
irs <- tempfile(fileext = ".csv")
write.csv(iris, irs)
mtc <- tempfile(fileext = ".csv")
write.csv(mtcars, mtc)
zzp <- tempfile(fileext = ".zip")
zip_create(c(irs, mtc), zzp, junk = TRUE)
zip_info(zzp)
zip_ls(zzp)
}
